Move ReactComponent out of Pods directory in EmbeddableApp
It doesn’t seem like a good practice to tell people to create their custom
ReactComponent directory inside of the Pods/ directory. The Pods/ directory
should be limited to CocoaPods installed content, and it’s not even
accepted across the board to [check this stuff into git][1]

Therefore, let’s tune the EmbeddableApp tutorial to instruct users to
create a ReactComponent directory at the root of their project. From there,
we’ll start the development server packager to source that directory as
a JS files root. This change also improves starting this packager server,
proxying the command through npm.

[1]: https://github.com/github/gitignore/blob/5883804e6ceb0867e88dc8ca1c7585fa61e40797/Objective-C.gitignore#L20-L26